+++ frontend/src/layout/Layout.tsx
@@ -1,0 +1,22 @@
+import React from 'react';
+import NavBar from './NavBar';
+import Footer from './Footer';
+import { Box } from '@mui/material';
+
+type LayoutProps = {
+  children: React.ReactNode;
+};
+
+const Layout: React.FC<LayoutProps> = ({ children }) => {
+  return (
+    <Box display="flex" flexDirection="column" minHeight="100vh">
+      <NavBar />
+      <Box component="main" flexGrow={1}>
+        {children}
+      </Box>
+      <Footer />
+    </Box>
+  );
+};
+
+export default Layout;
