+++ frontend/src/pages/Suburb/SuburbReportPage.tsx
@@ -2,14 +2,15 @@ import ActionButtonContainer from '@/components/containers/ActionButtonContainer
 import BannerContainer from '@/components/containers/BannerContainer';
 import BodyContainer from '@/components/containers/BodyContainer';
 import CardContainer from '@/components/containers/CardConatiner';
-import FooterContainer from '@/components/Footer';
-import NavbarContainer from '@/components/NavBar';
+import FooterContainer from '@/layout/Footer';
+import NavbarContainer from '@/layout/NavBar';
 import { TITLES } from '@/constants/titles';
 import type { AppDispatch, RootState } from '@/store';
 import { fetchSuburbReport, setSuburbId } from '@/store/slices/suburbSlice';
 import { Button, Typography } from '@mui/material';
 import React, { useEffect } from 'react';
 import { useDispatch, useSelector } from 'react-redux';
+import Layout from '@/layout/Layout';
 // import { useLocation } from 'react-router-dom';
 
 type SuburbReportPageProps = {
@@ -46,8 +47,8 @@ const SuburbReportPage: React.FC<SuburbReportPageProps> = () => {
   // const [stateCode, suburbName] = location.split('+');
 
   return (
-    <div>
-      <NavbarContainer></NavbarContainer>
+    <Layout>
+    
       <BannerContainer>
         {/* todo: replace with real banner content */}
         <Typography variant="h3" fontWeight={700}>
@@ -79,8 +80,8 @@ const SuburbReportPage: React.FC<SuburbReportPageProps> = () => {
           <Button>Export PDF</Button>
         </ActionButtonContainer>
       </BodyContainer>
-      <FooterContainer></FooterContainer>
-    </div>
+
+    </Layout>
   );
 };
